feat: Products controller and products service

+++ src/main/java/pl/oliwawyplywa/web/dto/products/ProductOptionDTO.java
@@ -2,18 +2,20 @@ package pl.oliwawyplywa.web.dto.products;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public class ProductOptionDTO {
 
     @JsonProperty("option_label")
     private String optionLabel;
 
     @JsonProperty("option_price")
-    private float optionPrice;
+    private BigDecimal optionPrice;
 
     public ProductOptionDTO() {
     }
 
-    public ProductOptionDTO(String optionLabel, float optionPrice) {
+    public ProductOptionDTO(String optionLabel, BigDecimal optionPrice) {
         this.optionLabel = optionLabel;
         this.optionPrice = optionPrice;
     }
@@ -26,11 +28,11 @@ public class ProductOptionDTO {
         this.optionLabel = optionLabel;
     }
 
-    public float getOptionPrice() {
+    public BigDecimal getOptionPrice() {
         return optionPrice;
     }
 
-    public void setOptionPrice(float optionPrice) {
+    public void setOptionPrice(BigDecimal optionPrice) {
         this.optionPrice = optionPrice;
     }
 }
