+++ src/main/java/pl/oliwawyplywa/web/dto/products/ResponseProductOptionDTO.java
@@ -2,6 +2,8 @@ package pl.oliwawyplywa.web.dto.products;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public class ResponseProductOptionDTO extends ProductOptionDTO {
 
     @JsonProperty("option_id")
@@ -11,7 +13,7 @@ public class ResponseProductOptionDTO extends ProductOptionDTO {
         super();
     }
 
-    public ResponseProductOptionDTO(String optionLabel, float optionPrice, int optionId) {
+    public ResponseProductOptionDTO(String optionLabel, BigDecimal optionPrice, int optionId) {
         super(optionLabel, optionPrice);
         this.optionId = optionId;
     }
